Extend vectorize module to generate data points for the var prediction task

+++ type4py/vectorize.py
@@ -15,9 +15,11 @@ W2V_VEC_LENGTH = 100
 AVAILABLE_TYPES_NUMBER = 1024
 
 class TokenIterator:
-    def __init__(self, param_df: pd.DataFrame, return_df: pd.DataFrame) -> None:
+    def __init__(self, param_df: pd.DataFrame, return_df: pd.DataFrame,
+                 var_df: pd.DataFrame) -> None:
         self.param_df = param_df
         self.return_df = return_df
+        self.var_df = var_df
 
     def __iter__(self):
         for return_expr_sentences in self.return_df['return_expr_str'][self.return_df['return_expr_str'] != '']:
@@ -32,13 +34,21 @@ class TokenIterator:
         for arg_names_sentences in self.return_df['arg_names_str'][self.return_df['arg_names_str'] != '']:
             yield arg_names_sentences.split()
 
+        for var_names_sentences in self.var_df['var_name'][self.var_df['var_name'].notnull()]:
+            yield var_names_sentences.split()
+
+        for var_occur_sentences in self.var_df['var_occur'][self.var_df['var_occur'] != '']:
+            yield var_occur_sentences.split()
+
 class W2VEmbedding:
     """
     Word2Vec embeddings for code tokens and identifiers
     """
-    def __init__(self, param_df: pd.DataFrame, return_df: pd.DataFrame, w2v_model_tk_path) -> None:
+    def __init__(self, param_df: pd.DataFrame, return_df: pd.DataFrame,
+                 var_df: pd.DataFrame, w2v_model_tk_path) -> None:
         self.param_df = param_df
         self.return_df = return_df
+        self.var_df = var_df
         self.w2v_model_tk_path = w2v_model_tk_path
 
     def train_model(self, corpus_iterator: TokenIterator, model_path_name: str) -> None:
@@ -48,28 +58,23 @@ class W2VEmbedding:
         :param model_path_name: path name of the output file
         """
 
-        cores = multiprocessing.cpu_count()
-
         w2v_model = Word2Vec(min_count=5,
                              window=5,
                              size=W2V_VEC_LENGTH,
-                             workers=cores-1)
+                             workers=multiprocessing.cpu_count())
 
         t = time()
-
         w2v_model.build_vocab(sentences=corpus_iterator)
-
         logger.info('Built W2V vocab in {} mins'.format(round((time() - t) / 60, 2)))
+        logger.info(f"W2V model's vocab size: {len(w2v_model.wv.vocab):,}")
 
         t = time()
-
         w2v_model.train(sentences=corpus_iterator,
                         total_examples=w2v_model.corpus_count,
                         epochs=20,
                         report_delay=1)
 
         logger.info('Built W2V model in {} mins'.format(round((time() - t) / 60, 2)))
-
         w2v_model.save(model_path_name)
 
     def train_token_model(self):
@@ -77,7 +82,7 @@ class W2VEmbedding:
         Trains a W2V model for tokens.
         """
 
-        self.train_model(TokenIterator(self.param_df, self.return_df), self.w2v_model_tk_path)
+        self.train_model(TokenIterator(self.param_df, self.return_df, self.var_df), self.w2v_model_tk_path)
 
 def vectorize_sequence(sequence: str, feature_length: int, w2v_model: Word2Vec) -> np.ndarray:
     """
@@ -100,12 +105,13 @@ class IdentifierSequence:
     """
     Vector representation of identifiers
     """
-    def __init__(self, identifiers_embd, arg_name, args_name, func_name):
+    def __init__(self, identifiers_embd, arg_name, args_name, func_name, var_name):
 
         self.identifiers_embd = identifiers_embd
         self.arg_name = arg_name
         self.args_name = args_name
         self.func_name = func_name
+        self.var_name = var_name
 
     def seq_length_param(self):
 
@@ -125,7 +131,12 @@ class IdentifierSequence:
             "padding": 10
         }
 
-    def generate_datapoint(self, seq_length):
+    def seq_length_var(self):
+        return {
+            'var_name': 10,
+        } # TODO: padding needed for combining with ret & param
+
+    def __gen_datapoint(self, seq_length):
         datapoint = np.zeros((sum(seq_length.values()), W2V_VEC_LENGTH))
         separator = np.ones(W2V_VEC_LENGTH)
 
@@ -148,31 +159,40 @@ class IdentifierSequence:
 
         return datapoint
 
-    def param_datapoint(self):
+    # def param_datapoint(self):
+    #     return self.__gen_datapoint(self.seq_length_param())
 
-        return self.generate_datapoint(self.seq_length_param())
+    # def return_datapoint(self):
+    #     return self.__gen_datapoint(self.seq_length_return())
 
-    def return_datapoint(self):
+    # def var_datapoint(self):
+    #     return self.__gen_datapoint(self.seq_length_var())
 
-        return self.generate_datapoint(self.seq_length_return())
-
+    def generate_datapoint(self):
+        if self.arg_name is not None and self.args_name is not None and self.func_name is None:
+            return self.__gen_datapoint(self.seq_length_param())
+        elif self.args_name is not None and self.func_name is not None:
+            return self.__gen_datapoint(self.seq_length_return())
+        elif self.var_name is not None:
+            return self.__gen_datapoint(self.seq_length_var())
 
 class TokenSequence:
     """
     Vector representation of code tokens
     """
-    def __init__(self, token_model, len_tk_seq, num_tokens_seq, args_usage, return_expr):
+    def __init__(self, token_model, len_tk_seq, num_tokens_seq, args_usage,
+                 return_expr, var_usage):
         self.token_model = token_model
         self.len_tk_seq = len_tk_seq
         self.num_tokens_seq = num_tokens_seq
         self.args_usage = args_usage
         self.return_expr = return_expr
+        self.var_usage = var_usage
 
     def param_datapoint(self):
         datapoint = np.zeros((self.num_tokens_seq*self.len_tk_seq, W2V_VEC_LENGTH))
         for i, w in enumerate(vectorize_sequence(self.args_usage, self.num_tokens_seq*self.len_tk_seq, self.token_model)):
             datapoint[i] = w
-
         return datapoint
 
     def return_datapoint(self):
@@ -183,8 +203,21 @@ class TokenSequence:
             for w in vectorize_sequence(self.return_expr, self.len_tk_seq, self.token_model):
                 datapoint[p] = w
                 p += 1
-
         return datapoint
+
+    def var_datapoint(self):
+        datapoint = np.zeros((self.num_tokens_seq*self.len_tk_seq, W2V_VEC_LENGTH))
+        for i, w in enumerate(vectorize_sequence(self.var_usage, self.num_tokens_seq*self.len_tk_seq, self.token_model)):
+            datapoint[i] = w
+        return datapoint
+
+    def generate_datapoint(self):
+        if self.args_usage is not None:
+            return self.param_datapoint()
+        elif self.return_expr is not None:
+            return self.return_datapoint()
+        elif self.var_usage is not None:
+            return self.var_datapoint()
 
 def process_datapoints(f_name, output_path, embedding_type, type, trans_func, cached_file: bool=False):
 
@@ -192,7 +225,7 @@ def process_datapoints(f_name, output_path, embedding_type, type, trans_func, ca
         df = pd.read_csv(f_name, na_filter=False)
         datapoints = df.apply(trans_func, axis=1)
 
-        datapoints_X = np.stack(datapoints.progress_apply(lambda x: x.return_datapoint() if 'ret' in type else x.param_datapoint()),
+        datapoints_X = np.stack(datapoints.progress_apply(lambda x: x.generate_datapoint()),
                                 axis=0)
         np.save(os.path.join(output_path, embedding_type + type + '_datapoints_x'), datapoints_X)
 
@@ -206,43 +239,52 @@ def type_vector(size, index):
     v[index] = 1
     return v
 
-def gen_aval_types_datapoints(df_params, df_ret, set_type, output_path, cached_file: bool=False):
+def gen_aval_types_datapoints(df_params, df_ret, df_var, set_type, output_path, cached_file: bool=False):
     """
     It generates data points for available types.
     """
 
     if not (os.path.exists(os.path.join(output_path, f'params_{set_type}_aval_types_dp.npy')) and os.path.exists(os.path.join(output_path,
-            f'ret_{set_type}_aval_types_dp.npy'))) or not cached_file:
+            f'ret_{set_type}_aval_types_dp.npy'))) and os.path.exists(os.path.join(output_path, f'var_{set_type}_aval_types_dp.npy')) \
+                 or not cached_file:
 
         df_params = pd.read_csv(df_params)
         df_ret = pd.read_csv(df_ret)
+        df_var = pd.read_csv(df_var)
 
         aval_types_params = np.stack(df_params.progress_apply(lambda row: type_vector(AVAILABLE_TYPES_NUMBER, row.param_aval_enc),
                                                     axis=1), axis=0)
         aval_types_ret = np.stack(df_ret.progress_apply(lambda row: type_vector(AVAILABLE_TYPES_NUMBER, row.ret_aval_enc),
                                             axis=1), axis=0)
+        aval_types_var = np.stack(df_var.progress_apply(lambda row: type_vector(AVAILABLE_TYPES_NUMBER, row.var_aval_enc),
+                                            axis=1), axis=0)
 
         np.save(os.path.join(output_path, f'params_{set_type}_aval_types_dp'), aval_types_params)
         np.save(os.path.join(output_path, f'ret_{set_type}_aval_types_dp'), aval_types_ret)
+        np.save(os.path.join(output_path, f'var_{set_type}_aval_types_dp'), aval_types_var)
 
-        return aval_types_params, aval_types_ret
+        return aval_types_params, aval_types_ret, aval_types_var
     else:
         logger.warn(f'file params_{set_type}_aval_types_dp.npy exists!')
         logger.warn(f'file ret_{set_type}_aval_types_dp.npy exists!')
-        return None, None
+        logger.warn(f'file var_{set_type}_aval_types_dp.npy exists!')
+        return None, None, None
 
-def gen_labels_vector(params_df: pd.DataFrame, returns_df: pd.DataFrame, set_type: str, output_path: str):
+def gen_labels_vector(params_df: pd.DataFrame, returns_df: pd.DataFrame, var_df: pd.DataFrame,
+                      set_type: str, output_path: str):
     """
     It generates a flattened labels vector
     """
 
     params_df = pd.read_csv(params_df)
     returns_df = pd.read_csv(returns_df)
+    var_df = pd.read_csv(var_df)
 
     np.save(os.path.join(output_path, f'params_{set_type}_dps_y_all'), params_df['arg_type_enc_all'].values)
     np.save(os.path.join(output_path, f'ret_{set_type}_dps_y_all'), returns_df['return_type_enc_all'].values)
+    np.save(os.path.join(output_path, f'var_{set_type}_dps_y_all'), var_df['var_type_enc_all'].values)
 
-    return params_df['arg_type_enc_all'].values, returns_df['return_type_enc_all'].values
+    return params_df['arg_type_enc_all'].values, returns_df['return_type_enc_all'].values, var_df['var_type_enc_all'].values
 
 def vectorize_args_ret(output_path: str):
     """
@@ -251,15 +293,17 @@ def vectorize_args_ret(output_path: str):
 
     param_df = pd.read_csv(os.path.join(output_path, "_ml_param_train.csv"), na_filter=False)
     return_df = pd.read_csv(os.path.join(output_path, "_ml_ret_train.csv"), na_filter=False)
+    var_df = pd.read_csv(os.path.join(output_path, "_ml_var_train.csv"), na_filter=False)
 
     logger.info(f"No. of parameters types in train set: {param_df.shape[0]:,}")
     logger.info(f"No. of returns types in train set: {return_df.shape[0]:,}")
+    logger.info(f"No. of variables types in train set: {var_df.shape[0]:,}")
 
-    embedder = W2VEmbedding(param_df, return_df, os.path.join(output_path, 'w2v_token_model.bin'))
+    embedder = W2VEmbedding(param_df, return_df, var_df,
+                            os.path.join(output_path, 'w2v_token_model.bin'))
     embedder.train_token_model()
 
     w2v_token_model = Word2Vec.load(os.path.join(output_path, 'w2v_token_model.bin'))
-    logger.info(f"W2V model's vocab size : {len(w2v_token_model.wv.vocab):,}")
 
     # Create dirs for vectors
     mk_dir_not_exist(os.path.join(output_path, "vectors"))
@@ -270,8 +314,10 @@ def vectorize_args_ret(output_path: str):
     tks_seq_len = (7, 3)
     vts_seq_len = (15, 5)
     # Vectorize functions' arguments
-    id_trans_func_param = lambda row: IdentifierSequence(w2v_token_model, row.arg_name, row.other_args, row.func_name)
-    token_trans_func_param = lambda row: TokenSequence(w2v_token_model, tks_seq_len[0], tks_seq_len[1], row.arg_occur, None)
+    id_trans_func_param = lambda row: IdentifierSequence(w2v_token_model, row.arg_name, row.other_args,
+                                                         row.func_name, None)
+    token_trans_func_param = lambda row: TokenSequence(w2v_token_model, tks_seq_len[0], tks_seq_len[1],
+                                                       row.arg_occur, None, None)
 
     # Identifiers
     logger.info("[arg][identifiers] Generating vectors")
@@ -298,8 +344,9 @@ def vectorize_args_ret(output_path: str):
                        'tokens_', 'param_test', token_trans_func_param)
 
     # Vectorize functions' return types
-    id_trans_func_ret = lambda row: IdentifierSequence(w2v_token_model, None, row.arg_names_str, row.name)
-    token_trans_func_ret = lambda row: TokenSequence(w2v_token_model, tks_seq_len[0], tks_seq_len[1], None, row.return_expr_str)
+    id_trans_func_ret = lambda row: IdentifierSequence(w2v_token_model, None, row.arg_names_str, row.name, None)
+    token_trans_func_ret = lambda row: TokenSequence(w2v_token_model, tks_seq_len[0], tks_seq_len[1], None,
+                                                     row.return_expr_str, None)
 
     # Identifiers
     logger.info("[ret][identifiers] Generating vectors")
@@ -325,28 +372,63 @@ def vectorize_args_ret(output_path: str):
                        os.path.join(output_path, "vectors", "test"),
                        'tokens_', 'ret_test', token_trans_func_ret)
 
+    # Vectorize variables types
+    id_trans_func_var = lambda row: IdentifierSequence(w2v_token_model, None, None, None, row.var_name)
+    token_trans_func_var = lambda row: TokenSequence(w2v_token_model, tks_seq_len[0], tks_seq_len[1], None,
+                                                     None, row.var_occur)
+
+    # Identifiers
+    logger.info("[var][identifiers] Generating vectors")
+    process_datapoints(os.path.join(output_path, "_ml_var_train.csv"),
+                       os.path.join(output_path, "vectors", "train"),
+                       'identifiers_', 'var_train', id_trans_func_var)
+    process_datapoints(os.path.join(output_path, "_ml_var_valid.csv"),
+                       os.path.join(output_path, "vectors", "valid"),
+                       'identifiers_', 'var_valid', id_trans_func_var)
+    process_datapoints(os.path.join(output_path, "_ml_var_test.csv"),
+                       os.path.join(output_path, "vectors", "test"),
+                       'identifiers_', 'var_test', id_trans_func_var)
+
+    # Tokens
+    logger.info("[var][code tokens] Generating vectors")
+    process_datapoints(os.path.join(output_path, "_ml_var_train.csv"),
+                       os.path.join(output_path, "vectors", "train"),
+                       'tokens_', 'var_train', token_trans_func_var)
+    process_datapoints(os.path.join(output_path, "_ml_var_valid.csv"),
+                       os.path.join(output_path, "vectors", "valid"),
+                       'tokens_', 'var_valid', token_trans_func_var)
+    process_datapoints(os.path.join(output_path, "_ml_var_test.csv"),
+                       os.path.join(output_path, "vectors", "test"),
+                       'tokens_', 'var_test', token_trans_func_var)
+    
     # Generate data points for visible type hints
     logger.info("[visible type hints] Generating vectors")
     gen_aval_types_datapoints(os.path.join(output_path, "_ml_param_train.csv"),
                               os.path.join(output_path, "_ml_ret_train.csv"),
+                              os.path.join(output_path, "_ml_var_train.csv"),
                               'train', os.path.join(output_path, "vectors", "train"))
     gen_aval_types_datapoints(os.path.join(output_path, "_ml_param_valid.csv"),
                               os.path.join(output_path, "_ml_ret_valid.csv"),
+                              os.path.join(output_path, "_ml_var_valid.csv"),
                               'valid', os.path.join(output_path, "vectors", "valid"))
     gen_aval_types_datapoints(os.path.join(output_path, "_ml_param_test.csv"),
                               os.path.join(output_path, "_ml_ret_test.csv"),
+                              os.path.join(output_path, "_ml_var_test.csv"),
                               'test', os.path.join(output_path, "vectors", "test"))
 
     # a flattened vector for labels
     logger.info("[true labels] Generating vectors")
     gen_labels_vector(os.path.join(output_path, "_ml_param_train.csv"),
                       os.path.join(output_path, "_ml_ret_train.csv"),
+                      os.path.join(output_path, "_ml_var_train.csv"),
                       'train', os.path.join(output_path, "vectors", "train"))
     gen_labels_vector(os.path.join(output_path, "_ml_param_valid.csv"),
                       os.path.join(output_path, "_ml_ret_valid.csv"),
+                      os.path.join(output_path, "_ml_var_valid.csv"),
                       'valid', os.path.join(output_path, "vectors", "valid"))
     gen_labels_vector(os.path.join(output_path, "_ml_param_test.csv"),
                       os.path.join(output_path, "_ml_ret_test.csv"),
+                      os.path.join(output_path, "_ml_var_test.csv"),
                       'test', os.path.join(output_path, "vectors", "test"))
